http-netty: accumulate the header block length as a long
Motivation

`HttpObjectDecoder` summed the header block length into an int and only
then compared it against `maxTotalHeaderFieldsLength`. With both header
limits configured close to `Integer.MAX_VALUE` the counter could wrap
negative and let the limit check pass, defeating the guard.

Modifications

Hold `currentHeadersBlockLength` as a long so the accumulation cannot
overflow. The limit check and its message are unchanged.

Result

The total header limit cannot be bypassed by wraparound. Reaching the
old overflow required ~2GB of headers in a single message, so there is
no practical test for it; the existing `totalHeadersLimitFor*` boundary
tests cover the unchanged check.

+++ servicetalk-http-netty/src/main/java/io/servicetalk/http/netty/HttpObjectDecoder.java
@@ -174,7 +174,11 @@ abstract class HttpObjectDecoder<T extends HttpMetaData> extends ByteToMessageDe
      * Line being parsed. Used for reporting location in exceptions
      */
     private int parsingLine;
-    private int currentHeadersBlockLength;
+    /**
+     * Running total of the header block being parsed, compared against {@link #maxTotalHeaderFieldsLength}. Held as a
+     * long so the accumulation cannot overflow and slip past the limit check.
+     */
+    private long currentHeadersBlockLength;
 
     /**
      * The internal state of {@link HttpObjectDecoder}.
